修复字段属性

+++ ItemEdit/ItemEdit/Models/profiles/Character.cs
@@ -203,7 +203,7 @@ namespace ItemEdit.Models.profiles
 
     public class Upd
     {
-        public int? StackObjectsCount { get; set; }
+        public float? StackObjectsCount { get; set; }
         public Tag Tag { get; set; }
         public Repairable Repairable { get; set; }
         public Firemode FireMode { get; set; }
